[core] Extract the .npy folder loader into heat/core/io/npy.py
`load_npy_from_path` moves out of io.py and is reimplemented on top of
`utils.load_from_folder`, which it shared almost line for line with
`load_csv_from_folder`. Behaviour is unchanged except that the files are
now read in sorted order.

Not registered with the format registry: it takes a directory, so there
is no extension for `load`/`save` to dispatch on -- same as before.
