+++ heat/core/io/io.py
@@ -43,7 +43,6 @@ __all__ = [
     "save",
     "supports_hdf5",
     "supports_netcdf",
-    "load_npy_from_path",
     "supports_zarr",
 ]
 
@@ -1341,62 +1340,6 @@ def save(
 
 DNDarray.save = lambda self, path, *args, **kwargs: save(self, path, *args, **kwargs)
 DNDarray.save.__doc__ = save.__doc__
-
-
-def load_npy_from_path(
-    path: str,
-    dtype: datatype = types.int32,
-    split: int = 0,
-    device: Optional[str] = None,
-    comm: Optional[Communication] = None,
-) -> DNDarray:
-    """
-    Loads multiple .npy files into one DNDarray which will be returned. The data will be concatenated along the split axis provided as input.
-
-    Parameters
-    ----------
-    path : str
-        Path to the directory in which .npy-files are located.
-    dtype : datatype, optional
-        Data type of the resulting array.
-    split : int
-        Along which axis the loaded arrays should be concatenated.
-    device : str, optional
-        The device id on which to place the data, defaults to globally set default device.
-    comm : Communication, optional
-        The communication to use for the data distribution, default is 'heat.MPI_WORLD'
-    """
-    if not isinstance(path, str):
-        raise TypeError(f"path must be str, not {type(path)}")
-    elif split is not None and not isinstance(split, int):
-        raise TypeError(f"split must be None or int, not {type(split)}")
-
-    process_number = MPI_WORLD.size
-    file_list = []
-    for file in os.listdir(path):
-        if fnmatch.fnmatch(file, "*.npy"):
-            file_list.append(file)
-    n_files = len(file_list)
-
-    if n_files == 0:
-        raise ValueError("No .npy Files were found")
-    if (n_files < process_number) and (process_number > 1):
-        raise RuntimeError("Number of processes can't exceed number of files")
-
-    rank = MPI_WORLD.rank
-    if rank < (n_files % process_number):
-        n_for_procs = n_files // process_number + 1
-        idx = rank * n_for_procs
-    else:
-        n_for_procs = n_files // process_number
-        idx = rank * n_for_procs + (n_files % process_number)
-    array_list = [np.load(path + "/" + element) for element in file_list[idx : idx + n_for_procs]]
-
-    larray = np.concatenate(array_list, split)
-    larray = torch.from_numpy(larray)
-
-    x = factories.array(larray, dtype=dtype, device=device, is_split=split, comm=comm)
-    return x
 
 
 try:
